test_requirements.txt
ansible-lint>=3.4.0
docker
jmespath
molecule>=3.3.0
molecule-docker
pytest==3.9.3
testinfra>=1.7.0
yamllint>=1.26.1
tox
molecule-robotframework
robotframework-seleniumlibrary
webdrivermanager